RMA-Demo/services/doc-processor/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
llama-parse==0.4.0
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.17.0
